refactor: removed default value to the createTelemetryMiddleware and datadog demo app adapted to the client v3
