+++ examples/datadog-express-apm/host-agent/package.json
@@ -1,11 +1,11 @@
 {
-  "name": "datadog",
+  "name": "my-app",
   "version": "1.0.0",
   "main": "index.js",
   "license": "MIT",
   "dependencies": {
     "cors": "^2.8.5",
     "dd-trace": "^5.4.0",
-    "express": "^4.19.2"
+    "express": "^4.21.2"
   }
 }
