+++ examples/datadog-express-apm/host-agent/utils/response.js
@@ -30,7 +30,7 @@ class ResponseHandler {
     agent.init().dogstatsd.increment(`Commercetools_Client_Response`, 1, {
       env: 'dev',
       status_code: statusCode,
-      http_method: req.method,
+
       success: true,
     })
 
@@ -43,7 +43,7 @@ class ResponseHandler {
    *
    * @description method to handle all error responses
    *
-   * @parma { response object }
+   * @param  response object
    * @param statusCode
    * @param message
    * @param data
@@ -63,7 +63,7 @@ class ResponseHandler {
     agent.init().dogstatsd.increment(`Commercetools_Client_Response`, 1, {
       env: 'dev',
       status_code: statusCode,
-      http_method: req.method,
+
       success: false,
     })
