+++ packages/ts-sdk-apm/src/helpers/newRelicHelper.ts
@@ -1,0 +1,6 @@
+import newrelic from 'newrelic'
+
+// record for newrelic
+export const recordNewRelic = (metric: number | newrelic.Metric): void => {
+  newrelic.recordMetric(`Commercetools/Client/Response/Total`, metric)
+}
