+++ packages/ts-sdk-apm/src/helpers/performanceHelper.ts
@@ -1,0 +1,1 @@
+export const time = () => performance.now()
