+++ packages/ts-sdk-apm/src/index.ts
@@ -1,4 +1,3 @@
 export * from '../types/types.d'
 export { default as createTelemetryMiddleware } from './apm'
 export { default as createTelemetryMiddlewareLegacy } from './apm-legacy'
-export * from './lib/agents';
