--- packages/ts-sdk-apm/src/lib/agents.ts
@@ -1,4 +1,0 @@
-import datadog from 'dd-trace'
-import newrelic from 'newrelic'
-
-export { datadog, newrelic }
